adds c++ runtime functions

+++ include/types.h
@@ -4,7 +4,7 @@
 #define TRUE 1
 
 typedef int bool_t;
-typedef long unsigned int size_t;
+typedef unsigned int size_t;
 
 typedef unsigned char byte_t;
